% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simul_withld}
\alias{simul_withld}
\title{Simulate Incident and subsequent trait for LD correlated SNPs}
\usage{
simul_withld(
  gpath,
  nind,
  h2_overall,
  h2_incidentonly,
  pd = 0,
  cv = 0.4,
  ev = 0.1,
  minfreq = 0.01,
  maxfreq = 0.49,
  pi = c(0.05, 0.05, 0.05)
)
}
\arguments{
\item{gpath}{path of the .gen file generated by hapgen}

\item{h2_overall}{total of the incident and subsequent events}

\item{h2_incidentonly}{proportion of heritability
affecting variance in outcome for incident event only}

\item{pd}{correlation between incident and subsequent event,
default value is 0}

\item{cv}{confounder variance for incident and subsequent traits,
default value is 0.4}

\item{ev}{error variance for incident and subsequent traits, default is 0.1}

\item{minfreq}{minimum value for allele frequency to simulate according to HWE
default value is 0.01}

\item{maxfreq}{maximum value for allele frequency to simulate according to HWE
default value is 0.49}

\item{pi}{proportion of SNPs that contribute to incident, subsequent and
both incident and subsequent traits, default value is 0.05 for all.}
}
\description{
This function allows you to simulate correlated incident and subsequent traits.
}
\examples{
simul_withld("chr19.hapgen_simul",0.5,0.25)
}
\keyword{LD,}
\keyword{correlated}
\keyword{simulate,}
\keyword{traits}
\keyword{with}
